requirements.txt
ascii_magic==2.3.0
imgkit==1.2.3
nudenet==3.0.8
NudeNetClassifier==2.1.1
python-telegram-bot==21.0.1
